remove seperate import tests
