--- azurestack/import_arm_security_rule_test.go
@@ -1,29 +1,0 @@
-package azurestack
-
-import (
-	"testing"
-
-	"github.com/hashicorp/terraform/helper/acctest"
-	"github.com/hashicorp/terraform/helper/resource"
-)
-
-func TestAccAzureStackNetworkSecurityRule_importBasic(t *testing.T) {
-	rInt := acctest.RandInt()
-	resourceName := "azurestack_network_security_rule.test"
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testCheckAzureStackNetworkSecurityRuleDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccAzureStackNetworkSecurityRule_basic(rInt, testLocation()),
-			},
-			{
-				ResourceName:      resourceName,
-				ImportState:       true,
-				ImportStateVerify: true,
-			},
-		},
-	})
-}
