--- azurestack/import_arm_virtual_network_test.go
@@ -1,31 +1,0 @@
-package azurestack
-
-import (
-	"testing"
-
-	"github.com/hashicorp/terraform/helper/acctest"
-	"github.com/hashicorp/terraform/helper/resource"
-)
-
-func TestAccAzureStackVirtualNetwork_importBasic(t *testing.T) {
-	resourceName := "azurestack_virtual_network.test"
-
-	ri := acctest.RandInt()
-	config := testAccAzureStackVirtualNetwork_basic(ri, testLocation())
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testCheckAzureStackVirtualNetworkDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: config,
-			},
-			{
-				ResourceName:      resourceName,
-				ImportState:       true,
-				ImportStateVerify: true,
-			},
-		},
-	})
-}
